repackage according to our screen session
