--- requirements.txt
@@ -1,2 +1,0 @@
-pdbpp
-deltachat>=1.70.0
